requirements.txt
Flask==1.0.2
mysql-connector-python==8.0.23
requests_oauthlib==0.5.0